refactor(drt): group selective and extensive searches in subpackages
